+++ CHANGELOG.md
@@ -6,6 +6,7 @@ versioning is done in a continuous fashion without worries of breaking changes.
 
 ## patches
 
+- `github-runners`: use known group for darwin dotfiles test runner 2026-03-28
 - `nvim`: grab permlink of highlighted changes for shared reference 2026-03-28
 - `zsh`: export nearest bin with above bin to paths for development 2026-03-23
 - `go`: use a latest release available to unstable packages release 2026-03-23